fix(hooks): Remove irrelevant cleanup
